+++ syndicate/core/groups/meta.py
@@ -1,13 +1,14 @@
 import json
 import os
-from functools import partial
 
 import click
 
 from syndicate.core.constants import (
     S3_BUCKET_ACL_LIST, API_GW_AUTHORIZER_TYPES, CUSTOM_AUTHORIZER_KEY,
     EC2_LAUNCH_TEMPLATE_SUPPORTED_IMDS_VERSIONS, EC2_LT_RESOURCE_TAGS,
+    FAILED_RETURN_CODE, OK_RETURN_CODE,
 )
+from syndicate.core.decorators import return_code_manager
 from syndicate.core.generators.deployment_resources import *
 from syndicate.core.generators.deployment_resources.api_gateway_generator import \
     ApiGatewayAuthorizerGenerator
@@ -16,7 +17,7 @@ from syndicate.core.generators.deployment_resources.ec2_launch_template_generato
     EC2LaunchTemplateGenerator
 from syndicate.core.generators.lambda_function import PROJECT_PATH_PARAM
 from syndicate.core.helper import (
-    OrderedGroup, OptionRequiredIf, validate_incompatible_options, check_tags,
+    OrderedGroup, OptionRequiredIf, check_tags,
     validate_authorizer_name_option, verbose_option, validate_api_gw_path,
     DictParamType, DeepDictParamType,
 )
@@ -29,6 +30,7 @@ dynamodb_type_param = click.Choice(['S', 'N', 'B'])
 
 
 @click.group(name=GENERATE_META_GROUP_NAME, cls=OrderedGroup)
+@return_code_manager
 @click.option('--project_path', nargs=1,
               help="Path to the project folder. Default value: the one "
                    "from the current config if it exists. "
@@ -39,17 +41,19 @@ def meta(ctx, project_path):
     """Generates deployment resources templates"""
     if not os.access(project_path, os.F_OK):
         click.echo(f"The provided path {project_path} doesn't exist")
-        return
+        return FAILED_RETURN_CODE
     elif not os.access(project_path, os.W_OK) or not os.access(project_path,
                                                                os.X_OK):
         click.echo(f"Incorrect permissions for the provided path "
                    f"'{project_path}'")
-        return
+        return FAILED_RETURN_CODE
     ctx.ensure_object(dict)
     ctx.obj[PROJECT_PATH_PARAM] = project_path
+    return OK_RETURN_CODE
 
 
 @meta.command(name='dax_cluster')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="Dax cluster name")
 @click.option('--node_type', required=True, type=str,
@@ -80,9 +84,11 @@ def dax_cluster(ctx, **kwargs):
     _generate(generator)
     click.echo(f'Dax cluster \'{kwargs["resource_name"]}\' was '
                f'successfully generated')
+    return OK_RETURN_CODE
 
 
 @meta.command(name='dynamodb')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="DynamoDB table name")
 @click.option('--hash_key_name', required=True, type=str,
@@ -114,9 +120,11 @@ def dynamodb(ctx, **kwargs):
     generator = DynamoDBGenerator(**kwargs)
     _generate(generator)
     click.echo(f"Table '{kwargs['resource_name']}' was added successfully!")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='dynamodb_global_index')
+@return_code_manager
 @click.option('--table_name', required=True, type=str,
               help="DynamoDB table name to add index to")
 @click.option('--name', required=True, type=str,
@@ -140,9 +148,11 @@ def dynamodb_global_index(ctx, **kwargs):
     generator = DynamoDBGlobalIndexGenerator(**kwargs)
     _generate(generator)
     click.echo(f"Global index '{kwargs['name']}' was added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='dynamodb_autoscaling')
+@return_code_manager
 @click.option('--table_name', type=str, required=True,
               help="DynamoDB table name to add autoscaling to")
 @click.option('--policy_name', type=str, required=True,
@@ -179,9 +189,11 @@ def dynamodb_autoscaling(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Autoscaling setting to table '{kwargs['table_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='s3_bucket')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="S3 bucket name", callback=check_bundle_bucket_name)
 @click.option('--location', type=ValidRegionParamType(),
@@ -223,9 +235,11 @@ def s3_bucket(ctx, **kwargs):
     _generate(generator)
     click.echo(f"S3 bucket '{kwargs['resource_name']}' was "
                f"added successfully!")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='api_gateway')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="Api gateway name")
 @click.option('--deploy_stage', required=True, type=str,
@@ -246,9 +260,11 @@ def api_gateway(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Api gateway '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='web_socket_api_gateway')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="Api gateway name")
 @click.option('--deploy_stage', required=True, type=str,
@@ -265,9 +281,11 @@ def web_socket_api_gateway(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Api gateway '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='api_gateway_authorizer')
+@return_code_manager
 @click.option('--api_name', required=True, type=str,
               help="Api gateway name to add index to")
 @click.option('--name', required=True, type=str,
@@ -292,9 +310,11 @@ def api_gateway_authorizer(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Authorizer '{kwargs['name']}' was added to API gateway "
                f"'{kwargs['api_name']}' successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='api_gateway_resource')
+@return_code_manager
 @click.option('--api_name', required=True, type=str,
               help="Api gateway name to add index to")
 @click.option('--path', required=True, callback=validate_api_gw_path,
@@ -312,9 +332,11 @@ def api_gateway_resource(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Resource '{kwargs['path']}' was added to API gateway "
                f"'{kwargs['api_name']}' successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='api_gateway_resource_method')
+@return_code_manager
 @click.option('--api_name', required=True, type=str,
               help="Api gateway name to add index to")
 @click.option('--path', required=True, callback=validate_api_gw_path,
@@ -360,9 +382,11 @@ def api_gateway_resource_method(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Method '{kwargs['method']}' was added to API gateway "
                f"resource '{kwargs['path']}' successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='iam_policy')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help='IAM policy name')
 @click.option('--policy_content', help='The path to JSON file with IAM policy '
@@ -386,9 +410,11 @@ def iam_policy(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Iam policy '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='iam_role')
+@return_code_manager
 @click.option('--resource_name', required=True, type=str,
               help="IAM role name")
 @click.option('--principal_service', required=True, type=str,
@@ -417,9 +443,11 @@ def iam_role(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Iam role '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='kinesis_stream')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Kinesis stream name")
 @click.option('--shard_count', type=int, required=True,
@@ -436,9 +464,11 @@ def kinesis_stream(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Kinesis stream '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='sns_topic')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="SNS topic name")
 @click.option('--region', type=ValidRegionParamType(allowed_all=True),
@@ -455,9 +485,11 @@ def sns_topic(ctx, **kwargs):
     _generate(generator)
     click.echo(f"SNS topic '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='step_function')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Step function name")
 @click.option('--iam_role', type=str, required=True,
@@ -474,9 +506,11 @@ def step_function(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Step function '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='step_function_activity')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Step function activity name")
 @click.option('--tags', type=DictParamType(), callback=check_tags,
@@ -491,9 +525,11 @@ def step_function_activity(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Step function activity '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='ec2_instance')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Instance name")
 @click.option('--key_name', type=str, required=True,
@@ -530,9 +566,11 @@ def ec2_instance(ctx, **kwargs):
     _generate(generator)
     click.echo(f"EC2 instance '{kwargs['resource_name']}' was added"
                f"successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='ec2_launch_template')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Launch template name")
 @click.option('--image_id', type=str, required=True,
@@ -585,9 +623,11 @@ def ec2_launch_template(ctx, **kwargs):
     _generate(generator)
     click.echo(f"ec2_launch_template '{kwargs['resource_name']}' was added "
                f"successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='sqs_queue')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="SQS queue name")
 @click.option('--region', type=ValidRegionParamType(),
@@ -643,9 +683,11 @@ def sqs_queue(ctx, **kwargs):
     _generate(generator)
     click.echo(f"SQS queue '{kwargs['resource_name']}' was added "
                f"successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="sns_application")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="The name of the sns application")
 @click.option('--platform', required=True,
@@ -666,9 +708,11 @@ def sns_application(ctx, **kwargs):
     _generate(generator)
     click.echo(f"SNS application '{kwargs['resource_name']}' was added "
                f"successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="cognito_user_pool")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Cognito user pool name")
 # @click.option('--region', type=ValidRegionParamType(), required=True,
@@ -708,9 +752,11 @@ def cognito_user_pool(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Cognito user pool '{kwargs['resource_name']}' was added "
                f"successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="cognito_federated_pool")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Cognito federated pool name")
 @click.option('--auth_role', type=str,
@@ -733,9 +779,11 @@ def cognito_federated_pool(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Cognito federated pool '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='batch_compenv')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Batch compute environment name")
 @click.option('--compute_environment_type',
@@ -797,9 +845,11 @@ def batch_compenv(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Batch compute environment '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name='batch_jobdef')
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help='Batch job definition name')
 @click.option('--job_definition_type', required=True,
@@ -823,9 +873,11 @@ def batch_jobdef(ctx, **kwargs):
     _generate(generator)
     click.echo(f'Batch job definition \'{kwargs["resource_name"]}\' was '
                f'added successfully')
+    return OK_RETURN_CODE
 
 
 @meta.command(name="batch_jobqueue")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Batch job queue name")
 @click.option('--state', type=click.Choice(["ENABLED", "DISABLED"]),
@@ -847,9 +899,11 @@ def batch_jobqueue(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Batch job queue '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="cloudwatch_alarm")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Cloudwatch alarm name")
 @click.option('--metric_name', type=str, required=True,
@@ -917,9 +971,11 @@ def cloudwatch_alarm(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Cloudwatch alarm '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="cloudwatch_event_rule")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="Cloudwatch event rule name")
 @click.option('--rule_type', required=True, help="Cloudwatch event rule type",
@@ -945,9 +1001,11 @@ def cloudwatch_event_rule(ctx, **kwargs):
     _generate(generator)
     click.echo(f"Cloudwatch event rule '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="eventbridge_rule")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="EventBridge rule name")
 @click.option('--rule_type', required=True, help="EventBridge rule type",
@@ -974,9 +1032,11 @@ def eventbridge_rule(ctx, **kwargs):
     _generate(generator)
     click.echo(f"EventBridge rule '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="documentdb_cluster")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="DocumentDB cluster name")
 @click.option('--master_username', type=str, required=True,
@@ -1005,9 +1065,11 @@ def documentdb_cluster(ctx, **kwargs):
     _generate(generator)
     click.echo(f"DocumentDB cluster '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 @meta.command(name="documentdb_instance")
+@return_code_manager
 @click.option('--resource_name', type=str, required=True,
               help="DocumentDB instance name")
 @click.option('--cluster_identifier', type=str, required=True,
@@ -1032,6 +1094,7 @@ def documentdb_instance(ctx, **kwargs):
     _generate(generator)
     click.echo(f"DocumentDB instance '{kwargs['resource_name']}' was "
                f"added successfully")
+    return OK_RETURN_CODE
 
 
 def _generate(generator: BaseConfigurationGenerator):