+++ syndicate/core/groups/tags.py
@@ -15,28 +15,33 @@
 """
 import click
 from syndicate.core.build.bundle_processor import load_deploy_output
+from syndicate.core.constants import OK_RETURN_CODE, FAILED_RETURN_CODE
+from syndicate.core.decorators import return_code_manager
 from syndicate.core.helper import verbose_option
 
 TAGS_GROUP_NAME = 'tags'
 
 
 @click.group(name=TAGS_GROUP_NAME)
+@return_code_manager
 @click.pass_context
 def tags(ctx):
     """Manage resources tags"""
     from syndicate.core import PROJECT_STATE, RESOURCES_PROVIDER
     if not PROJECT_STATE.latest_deploy:
         click.echo('No latest deploy')
-        raise click.Abort
+        raise click.Abort('No latest deploy')
     deploy_name = PROJECT_STATE.latest_deploy.get('deploy_name')
     bundle_name = PROJECT_STATE.latest_deploy.get('bundle_name')
     output = load_deploy_output(bundle_name, deploy_name)
     ctx.ensure_object(dict)
     ctx.obj['output'] = output
     ctx.obj['tags_resource'] = RESOURCES_PROVIDER.tags_api()
+    return OK_RETURN_CODE
 
 
 @tags.command(name='apply')
+@return_code_manager
 @verbose_option
 @click.pass_context
 def apply(ctx):
@@ -45,11 +50,14 @@ def apply(ctx):
     success = ctx.obj['tags_resource'].safe_apply_tags(output)
     if success:
         click.echo("Tags applied successfully")
+        return OK_RETURN_CODE
     else:
         click.echo("Tags applied with errors. More details in the log file")
+        return FAILED_RETURN_CODE
 
 
 @tags.command(name='remove')
+@return_code_manager
 @verbose_option
 @click.pass_context
 def remove(ctx):
@@ -58,5 +66,7 @@ def remove(ctx):
     success = ctx.obj['tags_resource'].safe_remove_tags(output)
     if success:
         click.echo("Tags removed successfully")
+        return OK_RETURN_CODE
     else:
         click.echo("Tags removed with errors. More details in the log file")
+        return FAILED_RETURN_CODE