+++ syndicate/core/project_state/project_state.py
@@ -27,7 +27,8 @@ import yaml
 from syndicate.commons.log_helper import get_logger
 from syndicate.core.constants import BUILD_ACTION, \
     DEPLOY_ACTION, UPDATE_ACTION, CLEAN_ACTION, PACKAGE_META_ACTION, \
-    ABORTED_STATUS, SUCCEEDED_STATUS, FAILED_STATUS
+    ABORTED_STATUS, SUCCEEDED_STATUS, FAILED_STATUS, FAILED_RETURN_CODE, \
+    OK_RETURN_CODE, ABORTED_RETURN_CODE
 from syndicate.core.constants import DATE_FORMAT_ISO_8601
 from syndicate.core.groups import RUNTIME_JAVA, RUNTIME_NODEJS, RUNTIME_PYTHON, \
     RUNTIME_SWAGGER_UI, RUNTIME_DOTNET, RUNTIME_APPSYNC
@@ -421,16 +422,27 @@ class ProjectState:
 
     def log_execution_event(self, **kwargs):
         operation = kwargs.get('operation')
-
         status = kwargs.get('status')
         rollback_on_error = kwargs.get('rollback_on_error')
-        if status not in [True, False, ABORTED_STATUS]:
-            kwargs.pop('status', None)
+        valid_statuses = {
+            FAILED_RETURN_CODE, OK_RETURN_CODE, ABORTED_RETURN_CODE
+        }
 
-        if operation in [DEPLOY_ACTION]:
+        if status not in valid_statuses:
+            kwargs.pop('status', None)
+        else:
+            status = {
+                OK_RETURN_CODE: True,
+                FAILED_RETURN_CODE: False,
+                ABORTED_RETURN_CODE: ABORTED_STATUS,
+                None: None
+            }.get(status)
+
+        if operation == DEPLOY_ACTION:
             params = kwargs.copy()
-            params.pop('operation')
-            params['is_succeeded'] = params.pop('status')
+            params.pop('operation', None)
+            params.pop('status', None)
+            params['is_succeeded'] = status
 
             if params['is_succeeded'] != ABORTED_STATUS:
                 if not (status is False and rollback_on_error is True):