+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support tags for `OAS V3` as part of the `oas_v3.json` file, using the `x-syndicate-openapi-tags` key
 - Improve the assembling of the Java lambda artifacts by removing the redundant original-<lambda_name>.jar file from the bundle
 - Remove the directory 'target' that contains temporary files after assembling lambda artifacts
+- Correct the return code of all the commands if they failed or were aborted
 
 # [1.15.0] - 2024-10-28
 - Added `--skip_tests` option to `build`, `test` and `assemble_java_mvn` commands to not run tests during or after